@@ -1,48 +1,43 @@
 {
-  "name": "hislogs",
+  "name": "hislogs.com",
   "private": false,
-  "description": "hislogs",
-  "version": "1.0.0",
+  "description": "blog",
+  "version": "2.0.0",
   "author": "sjsjsj1246 <sjsjsj1246@gmail.com>",
   "dependencies": {
-    "@fontsource/montserrat": "^4.2.2",
-    "@material-ui/core": "^4.11.3",
-    "@material-ui/icons": "^4.11.2",
-    "@mdx-js/mdx": "^1.6.22",
-    "@mdx-js/react": "^1.6.22",
-    "gatsby": "^2.32.3",
-    "gatsby-image": "^2.11.0",
-    "gatsby-plugin-advanced-sitemap": "^1.6.0",
-    "gatsby-plugin-dark-mode": "^1.1.2",
-    "gatsby-plugin-feed": "^2.1.0",
-    "gatsby-plugin-google-analytics": "^2.0.17",
-    "gatsby-plugin-gtag": "^1.0.13",
-    "gatsby-plugin-manifest": "^2.12.0",
-    "gatsby-plugin-offline": "^3.10.2",
-    "gatsby-plugin-react-helmet": "^3.10.0",
-    "gatsby-plugin-robots-txt": "^1.5.5",
-    "gatsby-plugin-sass": "^3.2.0",
-    "gatsby-plugin-sharp": "^2.14.3",
-    "gatsby-plugin-sitemap": "^4.8.0",
-    "gatsby-remark-autolink-headers": "^2.11.0",
-    "gatsby-remark-copy-linked-files": "^3.0.0",
-    "gatsby-remark-images": "^4.0.0",
-    "gatsby-remark-mathjax": "^1.0.0",
-    "gatsby-remark-prismjs": "^4.0.0",
-    "gatsby-remark-smartypants": "^3.0.0",
+    "@mui/icons-material": "^5.1.0",
+    "@mui/material": "^5.1.0",
+    "gatsby": "^3.12.1",
+    "gatsby-plugin-advanced-sitemap": "^2.0.0",
+    "gatsby-plugin-feed": "^3.13.0",
+    "gatsby-plugin-google-analytics": "^3.13.0",
+    "gatsby-plugin-image": "^1.14.0",
+    "gatsby-plugin-manifest": "^3.13.0",
+    "gatsby-plugin-offline": "^4.13.0",
+    "gatsby-plugin-react-helmet": "^4.13.0",
+    "gatsby-plugin-robots-txt": "^1.6.14",
+    "gatsby-plugin-sass": "^4.14.0",
+    "gatsby-plugin-sharp": "^3.13.0",
+    "gatsby-remark-autolink-headers": "^4.11.0",
+    "gatsby-remark-copy-linked-files": "^4.10.0",
+    "gatsby-remark-images": "^5.10.0",
+    "gatsby-remark-prismjs": "^5.10.0",
+    "gatsby-remark-responsive-iframe": "^4.10.0",
+    "gatsby-remark-smartypants": "^4.10.0",
     "gatsby-remark-table-of-contents": "^1.0.0",
-    "gatsby-source-filesystem": "^2.11.0",
-    "gatsby-theme-material-ui": "^1.0.13",
-    "gatsby-transformer-remark": "^2.16.1",
-    "gatsby-transformer-sharp": "^2.12.0",
-    "prismjs": "^1.23.0",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "react-helmet": "^6.1.0",
+    "gatsby-source-filesystem": "^3.13.0",
+    "gatsby-theme-material-ui": "^5.0.0",
+    "gatsby-transformer-remark": "^4.10.0",
+    "gatsby-transformer-sharp": "^3.13.0",
+    "prismjs": "^1.24.1",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
     "react-rotating-text": "^1.4.1",
-    "sass": "^1.32.8"
+    "sass": "^1.43.4",
+    "typeface-montserrat": "^1.1.13"
   },
   "devDependencies": {
+    "gh-pages": "^3.2.3",
     "prettier": "2.2.1"
   },
   "keywords": [
@@ -52,9 +47,10 @@
   ],
   "license": "0BSD",
   "scripts": {
-    "build": "gatsby build",
+    "build": "gatsby clean && gatsby build",
     "develop": "gatsby develop",
     "format": "prettier --write \"**/*.{js,jsx,ts,tsx,json,md}\"",
+    "deploy": "gatsby clean && gatsby build && gh-pages -d public",
     "start": "npm run develop",
     "serve": "gatsby serve",
     "clean": "gatsby clean",
