@@ -1,17 +1,16 @@
 import React, { useCallback } from 'react';
-import { IconButton, Tooltip } from '@material-ui/core';
-
-import EmailIcon from '@material-ui/icons/Email';
-import DescriptionIcon from '@material-ui/icons/Description';
-import PlayIcon from '@material-ui/icons/PlayArrowOutlined';
-import GitHubIcon from '@material-ui/icons/GitHub';
-import AndroidIcon from '@material-ui/icons/Android';
-import AppleIcon from '@material-ui/icons/Apple';
-import LinkedInIcon from '@material-ui/icons/LinkedIn';
+import { IconButton, Tooltip } from '@mui/material';
+import EmailIcon from '@mui/icons-material/Email';
+import DescriptionIcon from '@mui/icons-material/Description';
+import PlayIcon from '@mui/icons-material/PlayArrowOutlined';
+import GitHubIcon from '@mui/icons-material/GitHub';
+import AndroidIcon from '@mui/icons-material/Android';
+import AppleIcon from '@mui/icons-material/Apple';
+import LinkedInIcon from '@mui/icons-material/LinkedIn';
 
 import './style.scss';
 
-const IconButtonBar = ({ links = {} }) => {
+function IconButtonBar({ links = {} }) {
   const IconPicker = useCallback((icon) => {
     const props = { className: 'icon' };
     switch (icon) {
@@ -49,6 +48,6 @@ const IconButtonBar = ({ links = {} }) => {
       })}
     </>
   );
-};
+}
 
 export default IconButtonBar;
