@@ -1,6 +1,1 @@
-import '@fontsource/montserrat';
-import './src/styles/_reset.scss';
-import './src/styles/_colors.scss';
-import './src/styles/_global-style.scss';
-import './src/styles/_markdown-style.scss';
-import 'prismjs/themes/prism-tomorrow.css';
+require('typeface-montserrat');
