@@ -6,24 +6,6 @@ module.exports = {
   siteMetadata: metaConfig,
 
   plugins: [
-    `gatsby-plugin-advanced-sitemap`,
-    `gatsby-plugin-dark-mode`,
-    `gatsby-plugin-feed`,
-    `gatsby-plugin-offline`,
-    `gatsby-plugin-react-helmet`,
-    `gatsby-plugin-sharp`,
-    `gatsby-plugin-sass`,
-    `gatsby-theme-material-ui`,
-    `gatsby-plugin-sitemap`,
-    `gatsby-transformer-sharp`,
-    {
-      resolve: `gatsby-plugin-gtag`,
-      options: {
-        trackingId: metaConfig.ga,
-        head: false,
-        anonymize: true,
-      },
-    },
     {
       resolve: `gatsby-source-filesystem`,
       options: {
@@ -135,9 +117,17 @@ module.exports = {
           `gatsby-remark-autolink-headers`,
           `gatsby-remark-copy-linked-files`,
           `gatsby-remark-smartypants`,
-          `gatsby-remark-mathjax`,
         ],
       },
     },
+    `gatsby-theme-material-ui`,
+    `gatsby-transformer-sharp`,
+    `gatsby-plugin-advanced-sitemap`,
+    `gatsby-plugin-sharp`,
+    `gatsby-plugin-feed`,
+    `gatsby-plugin-image`,
+    `gatsby-plugin-offline`,
+    `gatsby-plugin-react-helmet`,
+    `gatsby-plugin-sass`,
   ],
 };
