@@ -1,13 +1,12 @@
 import React from 'react';
-
 import './style.scss';
 
-const PostContent = ({ html }) => (
-  <div className="post-content-wrapper">
+function PostContent({ html }) {
+  return (
     <div className="post-content">
       <div className="markdown" dangerouslySetInnerHTML={{ __html: html }} />
     </div>
-  </div>
-);
+  );
+}
 
 export default PostContent;
