@@ -1,43 +1,37 @@
-import React, { useState } from 'react';
+import React, { useCallback, useState } from 'react';
 import { graphql } from 'gatsby';
-import Layout from '../components/layout';
-import SEO from '../components/seo';
+import Layout from '../layout';
+import Seo from '../components/seo';
 import Bio from '../components/bio';
-import PostCardsColumn from '../components/post-cards-column';
 import Post from '../models/post';
-import Tabs from '../components/tabs';
 
-import { getSortedCategoriesByCount } from '../utils/helpers';
+import { getUniqueCategories } from '../utils/helpers';
+import PostTabs from '../components/post-tabs';
 
-export default ({ data }) => {
+function HomePage({ data }) {
   const posts = data.allMarkdownRemark.edges.map(({ node }) => new Post(node));
   const { author, language } = data.site.siteMetadata;
-  const categories = ['All', ...getSortedCategoriesByCount(posts)];
-  const [tabIndex, setTabIndex] = useState(0);
-
-  const onTabIndexChange = (e, value) => {
-    setTabIndex(value);
-  };
+  const categories = ['All', ...getUniqueCategories(posts)];
+  const featuredTabIndex = categories.findIndex((category) => category === 'featured');
+  const [tabIndex, setTabIndex] = useState(featuredTabIndex === -1 ? 0 : featuredTabIndex);
+  const onTabIndexChange = useCallback((e, value) => setTabIndex(value), []);
 
   return (
     <Layout>
-      <SEO title="Hislogs" />
+      <Seo title="Home" />
       <Bio author={author} language={language} />
-      <Tabs className={'tabs'} value={tabIndex} onChange={onTabIndexChange} tabs={categories} />
-      <PostCardsColumn
-        posts={
-          tabIndex === 0
-            ? posts.slice(0, 4)
-            : posts
-                .filter((post, index) => post.categories.includes(categories[tabIndex]))
-                .slice(0, 4)
-        }
-        moreUrl={`posts/${tabIndex === 0 ? '' : categories[tabIndex]}`}
+      <PostTabs
+        posts={posts}
+        onChange={onTabIndexChange}
+        tabs={categories}
+        tabIndex={tabIndex}
         showMoreButton
       />
     </Layout>
   );
-};
+}
+
+export default HomePage;
 
 export const pageQuery = graphql`
   query {
@@ -45,7 +39,7 @@ export const pageQuery = graphql`
       edges {
         node {
           id
-          excerpt(pruneLength: 200, truncate: true)
+          excerpt(pruneLength: 500, truncate: true)
           frontmatter {
             categories
             title
